+++ js/clusterKmeans.js
@@ -95,8 +95,10 @@ export class KmeansClusters {
      * This invalidates this object and all references to it.
      */
     free() {
-        this.results.delete();
-        this.results = null;
+        if (this.results !== null) {
+            this.results.delete();
+            this.results = null;
+        }
         return;
     }
 }
