+++ js/computePerCellQCFilters.js
@@ -106,8 +106,10 @@ export class PerCellQCFilters {
      * This invalidates this object and all references to it.
      */
     free() {
-        this.results.delete();
-        this.results = null;
+        if (this.results !== null) {
+            this.results.delete();
+            this.results = null;
+        }
         return;
     }
 }
