+++ js/computePerCellQCMetrics.js
@@ -62,8 +62,10 @@ export class PerCellQCMetrics {
      * This invalidates this object and all references to it.
      */
     free() {
-        this.results.delete();
-        this.results = null;
+        if (this.results !== null) {
+            this.results.delete();
+            this.results = null;
+        }
         return;
     }
 }
