+++ js/runPCA.js
@@ -68,9 +68,11 @@ export class PCAResults {
      * @return Frees the memory allocated on the Wasm heap for this object.
      * This invalidates this object and all references to it.
      */
-    free () {
-        this.results.delete();
-        this.results = null;
+    free() {
+        if (this.results !== null) {
+            this.results.delete();
+            this.results = null;
+        }
         return;
     }
 }
